grid_partial_plot

grid_partial_plot plots partial dependence plots from gbm.step models in ggplot

grid_partial_plot(x, vars, factors, ...)

Arguments

x

a gbm object

vars

a character vector of the variables you want to plot

factors

a character vector of those variables which are factors

...

additional options that you might want to send to gridExtra

Details

grid_partial_plot does basically the same job as gg_partial_plot but it has the capacity to plot factors alongside continuous variables. It does this by using gridExtra::grid.arrange. This might not be the most elegant or general solution, but it does work.